requirements.txt
# python-telegram-bot[job-queue]==20.7
# requests==2.31.0
# urllib3==1.26.18
# pytz==2023.3
# psycopg2-binary==2.9.9
# python-dotenv==1.0.0
python-telegram-bot[job-queue]==20.7
requests==2.31.0
pytz==2023.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0